Update README.md
The Contributors and List of Publications link was not active , so i, referenced the correct links in those categories. I was trying to create Publications. md page in R markdown , but Volesti logo image in google images is not opening. May be the link is broken in google images for Volesti Logo image. Hence , referenced it from the documentation